UPDATED : PATCH/DELETE method

+++ carts/views.py
@@ -10,7 +10,7 @@ from django.db.models  import Q
 from core.models       import TimeStampModel
 from core.views        import check_login
 from my_settings       import SECRET_KEY, ALGORITHM
-from users.models      import User
+from users.models      import User, Point
 from products.models   import Product, ProductEfficacy, ProductSummary, Efficacy
 from .models           import Cart
 
@@ -42,8 +42,9 @@ class CartView(View):
                 "productPrice"       : int(cart.product.price),
                 "thumbnail_image_url": cart.product.thumbnail_image_url,
             }for cart in carts]
+            point = Point.objects.get(user=user).point
 
-            return JsonResponse({"message":results},status=200)
+            return JsonResponse({"product":results,"point":point},status=200)
 
         except TypeError:
             return JsonResponse({"message":"TYPE_ERROR"},status=400)
@@ -53,26 +54,32 @@ class CartView(View):
     @check_login
     def patch(self, request):
         try: 
-            data    = json.loads(request.body)
-            user    = request.user
-            product = Product.objects.get(id=data["productID"])
-            
-            Cart.objects.filter(product_id=product, user_id=user).update(quantity = data["quantity"])
+            data       = json.loads(request.body)
+            user       = request.user
+            product_id = data["productID"]
+            product    = Product.objects.get(id=product_id)
 
+            if Cart.objects.filter(product_id=product, user_id=user).exists():
+                Cart.objects.filter(product_id=product, user_id=user).update(quantity = data["quantity"])
+            else: 
+                return JsonResponse({"message":f'{product_id}_IS_NOT_FOUND'},status=400)
+                
             return JsonResponse({"message":"UPDATE_COMPLETED"},status=200)
-
         except KeyError: 
             return JsonResponse({"message":"KEY_ERROR"},status=400)
 
     @check_login
     def delete(self, request):
         try: 
-            data    = json.loads(request.body)
-            user    = request.user
-            product = Product.objects.get(id=data["productID"])
+            data       = json.loads(request.body)
+            user       = request.user
+            product_id = data["productID"]
 
-            Cart.objects.filter(product_id=product, user_id=user).delete()
-
+            for product in product_id:
+                if Cart.objects.filter(product_id=product, user_id=user).exists():
+                    Cart.objects.filter(product_id=product, user_id=user).delete()
+                else: 
+                    return JsonResponse({"message":f'{product}_IS_NOT_FOUND'},status=204)
             return JsonResponse({"message":"DELETE_COMPLETED"},status=204)
         except KeyError: 
             return JsonResponse({"message":"KEY_ERROR"},status=400)
